* Added `query.WithPoolID()` execute option
